+++ CHANGELOG.md
@@ -16,6 +16,9 @@ List of changes that are finished but not yet released in any final version.
  - [PR-361](https://github.com/Cognifide/knotx/pull/361) - Improved Adapter / Knots creation - we do not create them with every request.
  - [PR-359](https://github.com/Cognifide/knotx/pull/359) - Fix for #355 - Knot.x responses with the response of a repository in case of response with status code different than 200.
  - [PR-369](https://github.com/Cognifide/knotx/pull/369) - Better support for SSL for Repository Connector
+ - [PR-371](https://github.com/Cognifide/knotx/pull/371) - Fixed debug logging of HTTP Repository Connector
+ - [PR-372](https://github.com/Cognifide/knotx/pull/372) - Added cache for compiled Handlebars snippets
+ - [PR-374](https://github.com/Cognifide/knotx/pull/374) - Enable keepAlive connection in http client options
 
 ## Version 1.1.2
  - [PR-318](https://github.com/Cognifide/knotx/pull/318) - Knot.x returns exit code `30` in case of missing config